feat: complete transition to HonoX
