+++ app/components/2022/Sponsors.tsx
@@ -1,7 +1,7 @@
-import { html } from 'hono/html'
+import type { FC } from 'hono/jsx'
 
-export const Sponsors = () => {
-  return html`
+export const Sponsors: FC = () => {
+  return (
     <section id="collaboradors" class="prose m-auto my-10 max-w-5xl px-4" style="scroll-margin-top: 84px">
       <h2 class="my-2 text-center font-heading text-3xl uppercase">Col·laboradors</h2>
       <p class="mb-4 text-justify">
@@ -9,7 +9,7 @@ export const Sponsors = () => {
         amb el projecte.
       </p>
       <div class="grid grid-cols-2 items-center justify-items-center gap-2 md:grid-cols-4">
-        <picture>
+        {/* <picture>
           <source srcset="/static/images/2023/esponsors/aperitius_tafaner.webp" type="image/webp" />
           <img
             src="/static/images/2023/esponsors/aperitius_tafaner.png"
@@ -306,8 +306,8 @@ export const Sponsors = () => {
             width="200"
             height="200"
           />
-        </picture>
+        </picture> */}
       </div>
     </section>
-  `
+  )
 }
