--- app/islands/counter.tsx
@@ -1,11 +1,0 @@
-import { useState } from 'hono/jsx'
-
-export default function Counter() {
-  const [count, setCount] = useState(0)
-  return (
-    <div>
-      <p>Count: {count}</p>
-      <button onClick={() => setCount(count + 1)}>Increment</button>
-    </div>
-  )
-}
