+++ package.json
@@ -5,13 +5,8 @@
   "type": "module",
   "scripts": {
     "build": "vite build --mode client && vite build",
-    "build:hono": "vite build",
-    "build:tailwind": "NODE_ENV=production pnpm dlx tailwindcss -i ./src/styles.css -o ./dist/static/styles.css --minify",
     "dev": "vite",
-    "dev:hono": "vite",
-    "dev:tailwind": "NODE_ENV=development pnpm dlx tailwindcss -i ./src/styles.css -o ./public/static/styles.css --watch",
-    "serve": "serve public",
-    "preview": "wrangler pages dev"
+    "preview": "wrangler pages dev ./dist"
   },
   "repository": {
     "type": "git",
@@ -31,24 +26,23 @@
   },
   "homepage": "https://fme.cat",
   "devDependencies": {
-    "@cloudflare/workers-types": "4.20240512.0",
+    "@cloudflare/workers-types": "4.20240529.0",
     "@hono/vite-cloudflare-pages": "0.4.0",
     "@hono/vite-dev-server": "0.12.1",
     "@hono/vite-ssg": "0.1.0",
     "concurrently": "8.2.2",
     "prettier": "3.2.5",
-    "prettier-plugin-tailwindcss": "0.5.14",
-    "serve": "14.2.3",
+    "prettier-plugin-tailwindcss": "0.6.0",
     "tailwindcss": "3.4.3",
-    "vite": "5.2.11",
-    "wrangler": "3.57.1"
+    "vite": "5.2.12",
+    "wrangler": "3.57.2"
   },
   "dependencies": {
     "@tailwindcss/typography": "0.5.13",
-    "autoprefixer": "^10.4.19",
-    "clsx": "^2.1.1",
-    "hono": "4.4.0",
-    "honox": "^0.1.18",
-    "postcss": "^8.4.38"
+    "autoprefixer": "10.4.19",
+    "clsx": "2.1.1",
+    "hono": "4.4.2",
+    "honox": "0.1.18",
+    "postcss": "8.4.38"
   }
 }
