--- src/types.ts
@@ -1,6 +1,0 @@
-export type PageProps = {
-  meta: {
-    title: string
-    description: string
-  }
-}
